+++ serde-query/src/lib.rs
@@ -11,16 +11,16 @@
 //!
 //! #[derive(DeserializeQuery)]
 //! struct Data {
-//!     #[query(".commit.author")]
-//!     commit_author: String,
+//!     #[query(".commit.authors.[0]")]
+//!     first_author: String,
 //!     #[query(".hash")]
 //!     hash_value: u64,
 //! }
 //!
 //! let document = serde_json::to_string(&serde_json::json!({
 //!     "commit": {
-//!         "author": "pandaman64",
-//!         "date": "2020/09/03",
+//!         "authors": ["Kou", "Kasumi", "Masaru"],
+//!         "date": "2020-09-10",
 //!     },
 //!     "hash": 0xabcd,
 //! }))?;
@@ -29,7 +29,7 @@
 //! // and convert the result to the desired type using `From`/`Into`.
 //! let data: Data = serde_json::from_str::<Query<Data>>(&document)?.into();
 //!
-//! assert_eq!(data.commit_author, "pandaman64");
+//! assert_eq!(data.first_author, "Kou");
 //! assert_eq!(data.hash_value, 0xabcd);
 //! # Ok(())
 //! # }
@@ -45,6 +45,10 @@
 //! `serde-query` currently supports the following syntax for accessing a part of the input.
 //!
 //! * `.field` for accessing a field with a name `field` of an object.
+//! * `.[index]` for accessing an array element at position `index`.
+//!
+//! Note that mixing field access and index access at the same position of a document
+//! is a compile error.
 //!
 //! [`DeserializeQuery`]: trait.DeserializeQuery.html
 //! [its derive macro]: derive.DeserializeQuery.html