+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Allow index queries.
+
 ## 0.1.0 - 2020-09-03
 
 ### Added